refactor(backend): Create licenses-core shared library
- Update parent backend pom to proper naming jar files sw360 prefixed
