+++ backend/licenses-core/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  ~ Copyright Siemens AG, 2013-2015. Part of the SW360 Portal Project.
+  ~ Copyright Siemens AG, 2016. Part of the SW360 Portal Project.
   ~
   ~ This program and the accompanying materials are made
   ~ available under the terms of the Eclipse Public License 2.0
@@ -9,8 +9,7 @@
   ~ SPDX-License-Identifier: EPL-2.0
   -->
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
@@ -19,11 +18,8 @@
         <version>18.99.1</version>
     </parent>
 
-    <artifactId>backend-licenses</artifactId>
-    <packaging>war</packaging>
-    <name>backend-licenses</name>
-    <build><finalName>licenses</finalName></build>
-
+    <artifactId>backend-licenses-core</artifactId>
+    <packaging>jar</packaging>
     <dependencies>
         <dependency>
             <groupId>org.eclipse.sw360</groupId>
@@ -31,9 +27,9 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.eclipse.sw360</groupId>
-            <artifactId>backend-licenses-core</artifactId>
-            <version>${project.version}</version>
+            <groupId>org.spdx</groupId>
+            <artifactId>tools-java</artifactId>
         </dependency>
+
     </dependencies>
 </project>
