+++ backend/licenses-core/src/main/java/org/eclipse/sw360/tools/OSADLObligationConnector.java
